Refactor dialog scripts
