+++ src/main.js
@@ -7,9 +7,9 @@ import './main/basicAuth';
 import { processDeepLink } from './main/deepLinks';
 import './main/updates';
 import { getMainWindow } from './main/mainWindow';
-import './main/dialogs/about';
-import './main/dialogs/screenshare';
-import './main/dialogs/update';
+import './main/aboutDialog';
+import './main/screenSharingDialog';
+import './main/updateDialog';
 import i18n from './i18n';
 
 export { default as dock } from './main/dock';
